feat: Add CommitBot code
Added the initial code for CommitBot. This code allows users to input their commit message and select a commit convention. It uses the git diff to provide context for the commit message and generates a commit message according to the selected convention using OpenAI's GPT-3 model. The generated commit message is displayed to the user. The user can also input a custom commit convention if desired.